Split up ErrorsView file into one for each component
