+++ src/browser/modules/Stream/CypherFrame/CypherFrame.tsx
@@ -31,7 +31,8 @@ import { SpinnerContainer, StyledStatsBarContainer } from '../styled'
 import { AsciiStatusbar, AsciiView } from './AsciiView'
 import { CancelView } from './CancelView'
 import { CodeStatusbar, CodeView } from './CodeView'
-import { ErrorsStatusbar, ErrorsView } from './ErrorsView/ErrorsView'
+import { ErrorsStatusbar } from './ErrorsView/ErrorsStatusbar'
+import { ErrorsView } from './ErrorsView/ErrorsView'
 import { PlanStatusbar, PlanView } from './PlanView'
 import { VisualizationConnectedBus } from './VisualizationView'
 import { WarningsStatusbar, WarningsView } from './WarningsView'
